fixed lint warnings and errors
