+++ packages/storybook/src/ui/molecules/icon/icon.stories.tsx
@@ -1,4 +1,5 @@
 import Icon from '@ltht-react/icon/src/molecules/icon'
+import EmotionIcon from '@ltht-react/icon/src/molecules/emotion-icon'
 import { Story } from '@storybook/react'
 
 export const AllPurposeIcon: Story = () => (
@@ -111,3 +112,22 @@ export const AllPurposeIcon: Story = () => (
     <br />
   </>
 )
+
+export const AllPurposeEmotionIcon: Story = () => (
+  <>
+    <div>chevron double - up</div>
+    <EmotionIcon type="chevron-double" size="1em" />
+    <br />
+    <div>chevron double - down</div>
+    <EmotionIcon type="chevron-double" size="1em" direction="down" />
+    <br />
+    <div>chevron double - left</div>
+    <EmotionIcon type="chevron-double" size="1em" direction="left" />
+    <br />
+    <div>chevron double - right</div>
+    <EmotionIcon type="chevron-double" size="1em" direction="right" />
+    <br />
+  </>
+)
+
+export default { title: 'UI/Molecules/Icon' }
