--- packages/storybook/src/ui/molecules/icon/index.stories.tsx
@@ -1,4 +1,0 @@
-export * from './icon.stories'
-export * from './emotion-icon.stories'
-
-export default { title: 'UI/Molecules/Icon' }
